unit tests for Item- and Stock- Service
Modified DeleteStockAsync method in StockService since it wasnt setup correctly. it would call the repository to delete the stock with that id even if there was no stock found with that id. There was a missing return function that would stop it from calling the repository if the id wasnt found